+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Validation for `defaultApproveNotificationEmailElementId`, `defaultClarificationNotificationEmailElementId` and `defaultDenyNotificationEmailElementId` properties
+- `approvalConfiguration` property to `Form`
 
 ## [0.19.11] - 2022-08-28